+++ Inc/C++Utilities/CppUtils.hpp
@@ -36,6 +36,7 @@
 #include <cstdarg>
 #include <stdarg.h>
 
+
 namespace chrono = std::chrono;
 namespace placeholders = std::placeholders;
 
@@ -67,9 +68,4 @@ using std::unordered_map;
 using std::vector;
 using std::queue;
 using std::map;
-
-template<class Type>
-concept Integral = is_integral<Type>::value;
-
-template<class Type>
-concept NotIntegral = !is_integral<Type>::value;
+using std::stringstream;
